+++ examples/src/example.js
@@ -38,9 +38,8 @@ const initHTML = `<br/>
 </center>
 <br/>
 <div><center><img src="${imageList[0]}" onclick="_.sendEvent('ImgClick')" contenteditable="false" height="170px"/></center></div>
-<div>
+<pre type="javascript"><code>const editor = ReactNative;</code><code>console.log(editor);</code></pre>
 <br/>Click the picture to switch<br/><br/>
-</div>
 `;
 
 const phizIcon = require('./assets/phiz.png');
@@ -88,8 +87,7 @@ class Example extends React.Component {
         Keyboard.removeListener('keyboardDidHide', this.onKeyHide);
     }
 
-    onKeyHide = () => {
-    };
+    onKeyHide = () => {};
 
     onKeyShow = () => {
         TextInput.State.currentlyFocusedInput() && this.setState({emojiVisible: false});
@@ -162,11 +160,11 @@ class Example extends React.Component {
 
     foreColor = () => {
         this.richText.current?.setForeColor('blue');
-    }
+    };
 
     hiliteColor = () => {
         this.richText.current?.setHiliteColor('red');
-    }
+    };
 
     insertHTML() {
         // this.richText.current?.insertHTML(
@@ -235,13 +233,30 @@ class Example extends React.Component {
         console.log('Paste:', data);
     };
 
+    /**
+     * @deprecated Android keyCode 229
+     * @param data
+     */
     handleKeyUp = data => {
         // console.log('KeyUp:', data);
     };
 
+    /**
+     * @deprecated Android keyCode 229
+     * @param data
+     */
     handleKeyDown = data => {
         // console.log('KeyDown:', data);
     };
+
+    /**
+     *
+     * @param data
+     * @param {string} inputType
+     */
+    onInput = ({data, inputType}) => {
+        // console.log(inputType, data)
+    }
 
     handleMessage = ({type, id, data}) => {
         let index = 0;
@@ -269,10 +284,10 @@ class Example extends React.Component {
         this.editorFocus = false;
     };
 
-    handleCursorPosition = (scrollY) => {
+    handleCursorPosition = scrollY => {
         // Positioning scroll bar
         this.scrollRef.current.scrollTo({y: scrollY - 30, animated: true});
-    }
+    };
 
     render() {
         let that = this;
@@ -281,7 +296,7 @@ class Example extends React.Component {
         const dark = theme === 'dark';
         return (
             <SafeAreaView style={[styles.container, dark && styles.darkBack]}>
-                <StatusBar barStyle={theme !== 'dark' ? 'dark-content' : 'light-content'}/>
+                <StatusBar barStyle={theme !== 'dark' ? 'dark-content' : 'light-content'} />
                 <InsertLinkModal
                     placeholderColor={placeholderColor}
                     color={color}
@@ -290,11 +305,15 @@ class Example extends React.Component {
                     ref={that.linkModal}
                 />
                 <View style={styles.nav}>
-                    <Button title={'HOME'} onPress={that.onHome}/>
-                    <Button title="Preview" onPress={that.save}/>
+                    <Button title={'HOME'} onPress={that.onHome} />
+                    <Button title="Preview" onPress={that.save} />
                 </View>
-                <ScrollView style={[styles.scroll, dark && styles.scrollDark]} keyboardDismissMode={'none'}
-                            ref={that.scrollRef} scrollEventThrottle={20}>
+                <ScrollView
+                    style={[styles.scroll, dark && styles.scrollDark]}
+                    keyboardDismissMode={'none'}
+                    ref={that.scrollRef}
+                    nestedScrollEnabled={true}
+                    scrollEventThrottle={20}>
                     <View style={[styles.topVi, dark && styles.darkBack]}>
                         <View style={styles.item}>
                             <Text style={{color}}>To: </Text>
@@ -315,8 +334,8 @@ class Example extends React.Component {
                             />
                         </View>
                         <View style={styles.item}>
-                            <Button title={theme} onPress={that.onTheme}/>
-                            <Button title={disabled ? 'enable' : 'disable'} onPress={that.onDisabled}/>
+                            <Button title={theme} onPress={that.onTheme} />
+                            <Button title={disabled ? 'enable' : 'disable'} onPress={that.onDisabled} />
                         </View>
                     </View>
                     <RichToolbar
@@ -346,6 +365,7 @@ class Example extends React.Component {
                         onPaste={that.handlePaste}
                         onKeyUp={that.handleKeyUp}
                         onKeyDown={that.handleKeyDown}
+                        onInput={that.onInput}
                         onMessage={that.handleMessage}
                         onFocus={that.handleFocus}
                         onBlur={that.handleBlur}
@@ -391,9 +411,7 @@ class Example extends React.Component {
                         ]} // default defaultActions
                         iconMap={{
                             insertEmoji: phizIcon,
-                            [actions.foreColor]: ({tintColor}) => (
-                                <Text style={[styles.tib, {color: 'blue'}]}>FC</Text>
-                            ),
+                            [actions.foreColor]: ({tintColor}) => <Text style={[styles.tib, {color: 'blue'}]}>FC</Text>,
                             [actions.hiliteColor]: ({tintColor}) => (
                                 <Text style={[styles.tib, {color: tintColor, backgroundColor: 'red'}]}>BC</Text>
                             ),
@@ -412,7 +430,7 @@ class Example extends React.Component {
                         foreColor={that.foreColor}
                         hiliteColor={that.hiliteColor}
                     />
-                    {emojiVisible && <EmojiView onSelect={that.insertEmoji}/>}
+                    {emojiVisible && <EmojiView onSelect={that.insertEmoji} />}
                 </KeyboardAvoidingView>
             </SafeAreaView>
         );
@@ -433,7 +451,7 @@ const styles = StyleSheet.create({
         minHeight: 300,
         flex: 1,
         borderBottomWidth: StyleSheet.hairlineWidth,
-        borderColor: '#e3e3e3'
+        borderColor: '#e3e3e3',
     },
     topVi: {
         backgroundColor: '#fafafa',